+++ actix-web/src/app.rs
@@ -21,8 +21,7 @@ use crate::{
     },
 };
 
-/// Application builder - structure that follows the builder pattern
-/// for building application instances.
+/// The top-level builder for an Actix Web application.
 pub struct App<T> {
     endpoint: T,
     services: Vec<Box<dyn AppServiceFactory>>,